+++ src/Codeception/Template/ModuleAcceptance/params.php
@@ -7,16 +7,20 @@
 
 declare(strict_types=1);
 
+use OxidEsales\Facts\Config\ConfigFile;
 use OxidEsales\Facts\Facts;
-use OxidEsales\Eshop\Core\ConfigFile;
-use OxidEsales\TestingLibrary\Services\Library\DatabaseDefaultsFileGenerator;
+use OxidEsales\Codeception\Module\Database\DatabaseDefaultsFileGenerator;
 
 $facts = new Facts();
 $selenium_server_port = getenv('SELENIUM_SERVER_PORT');
 $selenium_server_port = ($selenium_server_port) ? : '4444';
+$selenium_server_host = getenv('SELENIUM_SERVER_HOST');
+$selenium_server_host = ($selenium_server_host) ? : '127.0.0.1';
 $php = (getenv('PHPBIN')) ? : 'php';
-$cc_screen_shot_path = getenv('CC_SCREEN_SHOTS_PATH');
-$cc_screen_shot_path = ($cc_screen_shot_path) ? : '';
+$cc_screen_shot_url = getenv('CC_SCREEN_SHOTS_URL');
+$cc_screen_shot_url = ($cc_screen_shot_url) ? : '';
+$browser = getenv('BROWSER_NAME');
+$browser = ($browser) ? : 'firefox';
 
 return [
     'SHOP_URL' => $facts->getShopUrl(),
@@ -28,13 +32,21 @@ return [
     'DB_HOST' => $facts->getDatabaseHost(),
     'DB_PORT' => $facts->getDatabasePort(),
     'DUMP_PATH' => getTestDataDumpFilePath(),
+    'FIXTURES_PATH' => getTestFixtureSqlFilePath(),
     'MYSQL_CONFIG_PATH' => getMysqlConfigPath(),
     'SELENIUM_SERVER_PORT' => $selenium_server_port,
+    'SELENIUM_SERVER_HOST' => $selenium_server_host,
     'PHP_BIN' => $php,
-    'SCREEN_SHOT_URL' => $cc_screen_shot_path
+    'SCREEN_SHOT_URL' => $cc_screen_shot_url,
+    'BROWSER' => $browser
 ];
 
 function getTestDataDumpFilePath()
+{
+    return getShopTestPath() . '/{{testsDir}}/{{dataDir}}/generated/shop-dump.sql';
+}
+
+function getTestFixtureSqlFilePath()
 {
     return getShopTestPath() . '/{{testsDir}}/{{dataDir}}/dump.sql';
 }
@@ -64,8 +76,7 @@ function getShopTestPath()
 
 function getMysqlConfigPath()
 {
-    $facts = new Facts();
-    $configFile = new ConfigFile($facts->getSourcePath() . '/config.inc.php');
+    $configFile = new ConfigFile();
 
     $generator = new DatabaseDefaultsFileGenerator($configFile);