+++ src/Module/Oxideshop.php
@@ -7,38 +7,30 @@
 
 namespace OxidEsales\Codeception\Module;
 
-require_once __DIR__ . '/../../../../oxid-esales/testing-library/base.php';
-
-use Codeception\Exception\ElementNotFound;
 use Codeception\Lib\Interfaces\DependsOnModule;
 use Codeception\Module;
 use Codeception\Module\Db;
 use Codeception\Module\WebDriver;
+use Codeception\TestInterface;
 use Facebook\WebDriver\Exception\ElementNotVisibleException;
 use Facebook\WebDriver\Exception\NoSuchElementException;
 use OxidEsales\Facts\Facts;
 
 class Oxideshop extends Module implements DependsOnModule
 {
-    /**
-     * @var WebDriver
-     */
-    private $webDriver;
+    use CachingTrait;
+    use CommandTrait;
 
-    /**
-     * @var Db
-     */
-    private $db;
+    private WebDriver $webDriver;
+
+    private Db $database;
 
     /**
      * @var array
      */
     protected $config = ['screen_shot_url' => ''];
 
-    /**
-     * @return array
-     */
-    public function _depends()
+    public function _depends(): array
     {
         return [
             WebDriver::class => 'Codeception\Module\WebDriver is required',
@@ -46,82 +38,19 @@ class Oxideshop extends Module implements DependsOnModule
         ];
     }
 
-    /**
-     * @param WebDriver $driver
-     */
-    public function _inject(WebDriver $driver, Db $db)
+    public function _inject(WebDriver $driver, Db $database)
     {
         $this->webDriver = $driver;
-        $this->db = $db;
+        $this->database = $database;
     }
 
-    /**
-     * Reset context
-     */
-    public function _before(\Codeception\TestInterface $test)
+    public function _before(TestInterface $test)
     {
-        \OxidEsales\Codeception\Module\Context::resetActiveUser();
+        Context::resetActiveUser();
+        $this->cleanUpCompilationDirectory();
     }
 
-    /**
-     * Clear browser cache
-     */
-    public function clearShopCache()
-    {
-        $this->webDriver->_restart();
-    }
-
-    /**
-     * Clean up database
-     */
-    public function cleanUp()
-    {
-        $this->db->_beforeSuite([]);
-    }
-
-    /**
-     * Removes \n signs and it leading spaces from string. Keeps only single space in the ends of each row.
-     *
-     * @param string $line Not formatted string (with spaces and \n signs).
-     *
-     * @return string Formatted string with single spaces and no \n signs.
-     */
-    public function clearString(string $line)
-    {
-        return trim(preg_replace("/[ \t\r\n]+/", ' ', $line));
-    }
-
-    /**
-     * @param int $timeout
-     */
-    public function waitForAjax(int $timeout = 60)
-    {
-        $this->webDriver->waitForJS('y=(window.jQuery|$); return !y || y.active == 0;', $timeout);
-        $this->webDriver->wait(1);
-    }
-
-    /**
-     * @param int $timeout
-     */
-    public function waitForPageLoad(int $timeout = 60)
-    {
-        $this->waitForDocumentReadyState($timeout);
-        $this->waitForAjax($timeout);
-    }
-
-    /**
-     * @param int $timeout
-     */
-    public function waitForDocumentReadyState(int $timeout = 60)
-    {
-        $this->webDriver->waitForJs('return document.readyState == "complete"', $timeout);
-    }
-
-    /**
-     * @param \Codeception\TestInterface $test
-     * @param $fail
-     */
-    public function _failed(\Codeception\TestInterface $test, $fail)
+    public function _failed(TestInterface $test, $fail)
     {
         $report = $test->getMetadata()->getReports();
         if (isset($report['png']) && $this->config['screen_shot_url']) {
@@ -131,10 +60,49 @@ class Oxideshop extends Module implements DependsOnModule
         }
     }
 
+    public function clearShopCache(): void
+    {
+        $this->webDriver->_restart();
+    }
+
+    public function cleanUp(): void
+    {
+        $this->database->_beforeSuite([]);
+    }
+
+    /**
+     * Removes \n signs and it leading spaces from string. Keeps only single space in the ends of each row.
+     *
+     * @param string $line Not formatted string (with spaces and \n signs).
+     *
+     * @return string Formatted string with single spaces and no \n signs.
+     */
+    public function clearString(string $line): string
+    {
+        return trim(preg_replace("/[ \t\r\n]+/", ' ', $line));
+    }
+
+    public function waitForAjax(int $timeout = 60): void
+    {
+        $this->webDriver->waitForJS('y=(window.jQuery|$); return !y || y.active == 0;', $timeout);
+        $this->webDriver->wait(1);
+    }
+
+    public function waitForPageLoad(int $timeout = 60): void
+    {
+        $this->waitForDocumentReadyState($timeout);
+        $this->waitForAjax($timeout);
+    }
+
+    public function waitForDocumentReadyState(int $timeout = 60): void
+    {
+        $this->webDriver->waitForJs('return document.readyState == "complete"', $timeout);
+    }
+
     /**
      * Check if element exists on currently loaded page
      */
-    public function seePageHasElement($element)
+    public function seePageHasElement($element): bool
     {
         return count($this->getModule('WebDriver')->_findElements($element)) > 0;
     }
@@ -162,7 +130,6 @@ class Oxideshop extends Module implements DependsOnModule
 
     public function regenerateDatabaseViews(): void
     {
-        $vendorPath = (new Facts())->getVendorPath();
-        exec($vendorPath. '/bin/oe-eshop-db_views_generate');
+        $this->$this->processCommand((new Facts())->getVendorPath() . '/bin/oe-eshop-db_views_generate', []);
     }
 }