+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased] - unreleased
+## [2.1.0] - unreleased
 
 ### Added
 - SelectTheme module
+- ShopSetup module
+- CachingTrait
+- CommandTrait
+- DatabaseDefaultsFileGenerator class
 
 ## [2.0.0] - 2021-07-06